feat: refactored shadow terminal into its own crate
This was primarily to have a way of end to end testing Tattoy. It's
rather interesting that I already had what is probbaly the only usable
code on the internet for testing a fully-rendered terminal headlessly.

So not only do we now have a way to end to end test Tattoy, but the very
crate that allows it is itself a core part of Tattoy, namely the so-called
"Shadow Terminal". We are dogfooding.

The `shadow-terminal` crate is also written to be easy to use in
real-world use cases, such as Tattoy or say a terminal multiplexer.

This 2 uses (e2e test and multiplexing) mean that it would be a good
idea to publish this crate, once its had some good usage from me to iron
out any major issues.