feat: blending tattoy frames with alpha support
All frames can now be blended together. If a cell is blank and has a
true colour background then the alpha channel of the cell above is used
to blend into the background and text of the cell below.

Also adds the first tattoy that takes advantage of blending, namely a
scrollbar. The scrollbar tattoy is always enabled by default.