feat: minimap slide animation
The minimap now slides in and out when you hover over the right-most
side of the terminal. The "screen" surface is shown at the bottom of the
minimap and live updates.

Also refactored ShadowTerminal to always send the Screen surface.
Previously I was copying the bottom of the Scrollback in Tattoy to
always have a canonical copy of the terminal's live contents. I think
this approach a bit more efficient, but mostly just more intuitve.